refactor: 게임 관련 함수들을 game.js로 이동
