+++ src/utils/game.js
@@ -1,3 +1,18 @@
+import { Console, Random } from "@woowacourse/mission-utils";
+
+function generateComputerNumbers() {
+  const computer = [];
+
+  while (computer.length < 3) {
+    const number = Random.pickNumberInRange(1, 9);
+    if (!computer.includes(number)) {
+      computer.push(number);
+    }
+  }
+
+  return computer.join("");
+}
+
 function compareNumbers(computerNumbers, userInputNumber) {
   let strikeCount = 0;
   let ballCount = 0;
@@ -16,4 +31,16 @@ function compareNumbers(computerNumbers, userInputNumber) {
   return [strikeCount, ballCount];
 }
 
-export { compareNumbers };
+function printResultMessage(result) {
+  const [strikeCount, ballCount] = result;
+  let resultMessage = "";
+
+  if (strikeCount === 0 && ballCount === 0) resultMessage = "낫싱";
+  else if (strikeCount === 0) resultMessage = `${ballCount}볼`;
+  else if (ballCount === 0) resultMessage = `${strikeCount}스트라이크`;
+  else resultMessage = `${ballCount}볼 ${strikeCount}스트라이크`;
+
+  Console.print(resultMessage);
+}
+
+export { generateComputerNumbers, compareNumbers, printResultMessage };
